Add Booking Page booking details Showing page and also make many diffrent type of page

+++ back-end/back-end/Controllers/OrdersController.cs
@@ -38,6 +38,8 @@ namespace back_end.Controllers
             {
                 orders = await _context.Orders
                .Include(r => r.Booking)
+               .Include(p=>p.Booking.Product)
+               .Include(p => p.Booking.Product.Brand)
                .Include(s => s.Staff)
                .Where(r => r.OrderStatus == OrderStatus.Delivered || r.OrderStatus == OrderStatus.Rejected)
                .ToListAsync();
@@ -46,6 +48,8 @@ namespace back_end.Controllers
             {
                 orders = await _context.Orders
                 .Include(r => r.Booking)
+                .Include(p => p.Booking.Product)
+                .Include(p => p.Booking.Product.Brand)
                 .Include(s => s.Staff)
                 .Where(r => r.OrderStatus == OrderStatus.Placed || r.OrderStatus == OrderStatus.Confirmed || r.OrderStatus == OrderStatus.OnTheWay)
                 .ToListAsync();
@@ -106,10 +110,7 @@ namespace back_end.Controllers
             else
             {
                 order.IsStaffAccepted = false;
-                
                 order.UpdatedBy = Guid.Parse(userId);
-                
-                return Ok("Order Not Accepted");
             }
             _context.Orders.Update(order);
             await _context.SaveChangesAsync();
@@ -139,32 +140,63 @@ namespace back_end.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutOrder(Guid id, Order order)
+        public async Task<ActionResult<Order>> PutOrder(Guid id,OrderDTO orderDTO)
         {
-            if (id != order.OrderId)
+            if (id.ToString() != orderDTO.OrderId)
             {
                 return BadRequest();
             }
-
-            _context.Entry(order).State = EntityState.Modified;
-
-            try
+            var userIdClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+           
+            if (userIdClaim == null)
             {
-                await _context.SaveChangesAsync();
+                return Unauthorized("User ID not found in token.");
             }
-            catch (DbUpdateConcurrencyException)
+            var userId = userIdClaim.Value;
+
+            if (_context.Orders == null)
             {
-                if (!OrderExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
+                return Problem("Entity set 'ApplicationDbContext.Orders'  is null.");
+            }
+            if (!Enum.TryParse(typeof(PaymentType), orderDTO.PaymentType.ToString(), out var paymenttype))
+            {
+                return BadRequest("Invalid Payment Type.");
+            }
+            if (!Enum.TryParse(typeof(PaymentStatus), orderDTO.PaymentStatus.ToString(), out var paymentstatus))
+            {
+                return BadRequest("Invalid Payment status.");
             }
 
-            return NoContent();
+            if (!Enum.TryParse(typeof(OrderStatus), orderDTO.OrderStatus.ToString(), out var orderStatus))
+            {
+                return BadRequest("Invalid Order status.");
+            }
+            var existingOrder = await _context.Orders.Include(b=>b.Booking).FirstOrDefaultAsync(r=>r.OrderId == id);
+
+            if (existingOrder == null)
+            {
+                BadRequest("No Any Order Will Be Found");
+            }
+            existingOrder.OrderStatus =  (OrderStatus)orderStatus;
+            existingOrder.StaffId = Guid.Parse(orderDTO.StaffId);
+            if(orderDTO.IsStaffAccepted != null)
+            {
+                existingOrder.IsStaffAccepted = true;
+            }
+            if(orderDTO.OrderStatus == "Delivered")
+            {
+                existingOrder.DeliveryDate = DateTime.UtcNow;
+                if((PaymentType)paymenttype == PaymentType.COD)
+                {
+                    existingOrder.PaymentStatus = PaymentStatus.Success;
+                    existingOrder.Booking.PaymentDate = DateTime.UtcNow;
+                    existingOrder.Booking.PaymentId = "Cash Payment";
+                    existingOrder.Booking.PaymentStatus = PaymentStatus.Success;
+                }
+            }
+            await _context.SaveChangesAsync();
+            return existingOrder;
+
         }
 
         [HttpPost]
@@ -178,9 +210,9 @@ namespace back_end.Controllers
             var userId = userIdClaim.Value;
 
             if (_context.Orders == null)
-          {
+            {
               return Problem("Entity set 'ApplicationDbContext.Orders'  is null.");
-          }
+            }
             if (!Enum.TryParse(typeof(PaymentType), orderDTO.PaymentType.ToString(), out var paymenttype))
             {
                 return BadRequest("Invalid Payment Type.");
@@ -188,6 +220,19 @@ namespace back_end.Controllers
             if (!Enum.TryParse(typeof(PaymentStatus), orderDTO.PaymentStatus.ToString(), out var paymentstatus))
             {
                 return BadRequest("Invalid Payment status.");
+            }
+            var product = await _context.Products.FirstOrDefaultAsync(p => p.ProductId.ToString() == orderDTO.ProductId);
+            if (product == null)
+            {
+                return BadRequest("Product not available.");
+            }
+            if (!int.TryParse(product.Quantity, out int quantity))
+            {
+                return BadRequest("Quantity type is not correct.");
+            }
+            if (quantity <= 0)
+            {
+                return BadRequest("Stock is empty. Please try again later.");
             }
             var order = new Order
             {
@@ -205,6 +250,7 @@ namespace back_end.Controllers
                 StaffId=Guid.Parse(orderDTO.StaffId),
                 OrderStatus = OrderStatus.Placed,
             };
+            product.Quantity = (quantity - 1).ToString();
 
             var booking = await _context.Bookings.FirstOrDefaultAsync(r => r.BookingId.ToString() == orderDTO.BookingId);
             if(booking != null)
@@ -213,6 +259,7 @@ namespace back_end.Controllers
             }
 
             _context.Bookings.Update(booking);
+            _context.Products.Update(product);
             _context.Orders.Add(order);
             await _context.SaveChangesAsync();
 
@@ -236,12 +283,26 @@ namespace back_end.Controllers
                 return NotFound();
             }
             var order = await _context.Orders.FindAsync(id);
+            var booking = await _context.Bookings.FirstOrDefaultAsync(r => r.BookingId == order.BookingId);
+            booking.Status = BookingStatus.Pending;
+            var product = await _context.Products.FirstOrDefaultAsync(p => p.ProductId == booking.ProductID);
+            if (product == null)
+            {
+                return BadRequest("Product not available.");
+            }
+            if (!int.TryParse(product.Quantity, out int quantity))
+            {
+                return BadRequest("Quantity type is not correct.");
+            }
             if (order == null)
             {
                 return BadRequest();
             }
+            product.Quantity = (quantity + 1).ToString();
 
             _context.Orders.Remove(order);
+            _context.Bookings.Update(booking);
+            _context.Products.Update(product);
             await _context.SaveChangesAsync();
 
             return Ok("Data Successfully Deleted");
